+++ src/ripple/test/jtx/Env.h
@@ -23,6 +23,7 @@
 #include <ripple/test/jtx/Account.h>
 #include <ripple/test/jtx/amounts.h>
 #include <ripple/test/jtx/JTx.h>
+#include <ripple/test/jtx/require.h>
 #include <ripple/test/jtx/tags.h>
 
 #include <ripple/app/ledger/Ledger.h>
