+++ src/ripple/test/jtx/requires.h
@@ -17,32 +17,23 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_TEST_JTX_H_INCLUDED
-#define RIPPLE_TEST_JTX_H_INCLUDED
+#ifndef RIPPLE_TEST_JTX_REQUIRES_H_INCLUDED
+#define RIPPLE_TEST_JTX_REQUIRES_H_INCLUDED
 
-// Convenience header that includes everything
+#include <functional>
+#include <vector>
 
-#include <ripple/test/jtx/Account.h>
-#include <ripple/test/jtx/amounts.h>
-#include <ripple/test/jtx/any.h>
-//#include <ripple/test/jtx/balance.h>
-#include <ripple/test/jtx/Env.h>
-#include <ripple/test/jtx/fee.h>
-#include <ripple/test/jtx/json.h>
-#include <ripple/test/jtx/JTx.h>
-#include <ripple/test/jtx/multisign.h>
-#include <ripple/test/jtx/owners.h>
-#include <ripple/test/jtx/paths.h>
-#include <ripple/test/jtx/require.h>
-#include <ripple/test/jtx/requires.h>
-#include <ripple/test/jtx/sendmax.h>
-#include <ripple/test/jtx/seq.h>
-#include <ripple/test/jtx/sig.h>
-#include <ripple/test/jtx/tags.h>
-#include <ripple/test/jtx/ter.h>
-#include <ripple/test/jtx/ticket.h>
-#include <ripple/test/jtx/txflags.h>
-#include <ripple/test/jtx/utility.h>
+namespace ripple {
+namespace test {
+namespace jtx {
+
+class Env;
+
+using require_t = std::function<void(Env const&)>;
+using requires_t = std::vector<require_t>;
+
+} // jtx
+} // test
+} // ripple
 
 #endif
-
